+++ src/main/java/com/zyyglxt/service/impl/HospServiceImpl.java
@@ -3,12 +3,18 @@ package com.zyyglxt.service.impl;
 import com.zyyglxt.dao.HospDOMapper;
 import com.zyyglxt.dataobject.HospDO;
 import com.zyyglxt.dataobject.HospDOKey;
+import com.zyyglxt.error.BusinessException;
+import com.zyyglxt.error.EmBusinessError;
 import com.zyyglxt.service.IHospService;
+import com.zyyglxt.validator.ValidatorImpl;
+import com.zyyglxt.validator.ValidatorResult;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import javax.annotation.Resource;
+import java.util.Date;
 import java.util.List;
+import java.util.UUID;
 
 /**
  * @author qjc
@@ -19,19 +25,34 @@ import java.util.List;
 public class HospServiceImpl implements IHospService {
     @Resource
     HospDOMapper hospDOMapper;
+    @Autowired
+    private ValidatorImpl validator;
 
     @Override
     public int addHosp(HospDO hospDO) {
+        ValidatorResult result = validator.validate(hospDO);
+        if(result.isHasErrors()){
+            throw new BusinessException(result.getErrMsg(), EmBusinessError.PARAMETER_VALIDATION_ERROR);
+        }
+        hospDO.setItemcreateat(new Date());
         return hospDOMapper.insertSelective(hospDO);
     }
 
     @Override
     public int updateHosp(HospDO hospDO) {
+        ValidatorResult result = validator.validate(hospDO);
+        if(result.isHasErrors()){
+            throw new BusinessException(result.getErrMsg(), EmBusinessError.PARAMETER_VALIDATION_ERROR);
+        }
         return hospDOMapper.updateByPrimaryKeySelective(hospDO);
     }
 
     @Override
     public int deleteHosp(HospDOKey hospDOKey) {
+        ValidatorResult result = validator.validate(hospDOKey);
+        if(result.isHasErrors()){
+            throw new BusinessException(result.getErrMsg(), EmBusinessError.PARAMETER_VALIDATION_ERROR);
+        }
         return hospDOMapper.deleteByPrimaryKey(hospDOKey);
     }
 
@@ -45,12 +66,20 @@ public class HospServiceImpl implements IHospService {
      */
     @Override
     public List<HospDO> searchHosp(String keyWord) {
+        if(keyWord.isEmpty()){
+            throw new BusinessException("关键字不能为空", EmBusinessError.PARAMETER_VALIDATION_ERROR);
+        }
         return hospDOMapper.searchHosp(keyWord);
     }
 
     @Override
     public List<HospDO> top5Hosp() {
         return hospDOMapper.top5Hosp();
+    }
+
+    @Override
+    public HospDO selectHospByItemCode(String itemCode) {
+        return hospDOMapper.selectHospByItemCode(itemCode);
     }
 
 }
