+++ src/main/java/com/zyyglxt/service/impl/SpecialtyServiceImpl.java
@@ -6,11 +6,20 @@ import com.zyyglxt.dataobject.HospSpecialtyRefDO;
 import com.zyyglxt.dataobject.HospSpecialtyRefDOKey;
 import com.zyyglxt.dataobject.SpecialtyDO;
 import com.zyyglxt.dataobject.SpecialtyDOKey;
+import com.zyyglxt.dto.SpecialtyDto;
+import com.zyyglxt.error.BusinessException;
+import com.zyyglxt.error.EmBusinessError;
 import com.zyyglxt.service.ISpecialtyService;
+import com.zyyglxt.validator.ValidatorImpl;
+import com.zyyglxt.validator.ValidatorResult;
+import org.springframework.beans.BeanUtils;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import javax.annotation.Resource;
+import java.util.Date;
 import java.util.List;
+import java.util.UUID;
 
 /**
  * @author qjc
@@ -21,29 +30,69 @@ import java.util.List;
 public class SpecialtyServiceImpl implements ISpecialtyService {
 
     @Resource
-    SpecialtyDOMapper specialtyDOMapper;
+    private SpecialtyDOMapper specialtyDOMapper;
     @Resource
-    HospSpecialtyRefDOMapper hospSpecialtyRefDOMapper;
+    private HospSpecialtyRefDOMapper hospSpecialtyRefDOMapper;
+    @Autowired
+    private ValidatorImpl validator;
 
-    /*在控制层将数据封装好再传进service层*/
+    private SpecialtyDO specialtyDO = new SpecialtyDO();
+
+    private HospSpecialtyRefDO hospSpecialtyRefDO = new HospSpecialtyRefDO();
 
     /*增加科室，同时也将科室与医院记录插入*/
     @Override
-    public void addSpecialty(SpecialtyDO specialtyDO, HospSpecialtyRefDO hospSpecialtyRefDO) {
+    public void addSpecialty(SpecialtyDto specialtyDto) {
+        ValidatorResult result = validator.validate(specialtyDto);
+        if(result.isHasErrors()){
+            throw new BusinessException(result.getErrMsg(), EmBusinessError.PARAMETER_VALIDATION_ERROR);
+        }
+        BeanUtils.copyProperties(specialtyDto,specialtyDO);
+        specialtyDO.setItemcreateat(new Date());
+
+        hospSpecialtyRefDO.setItemcode(UUID.randomUUID().toString());
+        hospSpecialtyRefDO.setHospitalCode(specialtyDto.getHospitalCode());
+        hospSpecialtyRefDO.setSpecialtyCode(specialtyDto.getHospitalCode());
+        hospSpecialtyRefDO.setCreater(specialtyDto.getCreater());
+        hospSpecialtyRefDO.setItemcreateat(new Date());
+        hospSpecialtyRefDO.setUpdater(specialtyDto.getUpdater());
+
         specialtyDOMapper.insertSelective(specialtyDO);
         hospSpecialtyRefDOMapper.insertSelective(hospSpecialtyRefDO);
     }
 
     /*更新科室信息，同步更新医院科室关系表*/
     @Override
-    public void updateSpecialty(SpecialtyDO specialtyDO, HospSpecialtyRefDO hospSpecialtyRefDO) {
+    public void updateSpecialty(SpecialtyDto specialtyDto) {
+        ValidatorResult result = validator.validate(specialtyDto);
+        if(result.isHasErrors()){
+            throw new BusinessException(result.getErrMsg(), EmBusinessError.PARAMETER_VALIDATION_ERROR);
+        }
+        specialtyDO = specialtyDto;
+
+        hospSpecialtyRefDO.setHospitalCode(specialtyDto.getHospitalCode());
+        hospSpecialtyRefDO.setSpecialtyCode(specialtyDto.getHospitalCode());
+        hospSpecialtyRefDO.setCreater(specialtyDto.getCreater());
+        hospSpecialtyRefDO.setItemcreateat(specialtyDto.getItemcreateat());
+        hospSpecialtyRefDO.setUpdater(specialtyDto.getUpdater());
+        hospSpecialtyRefDO.setItemupdateat(specialtyDto.getItemupdateat());
+
         specialtyDOMapper.updateByPrimaryKeySelective(specialtyDO);
         hospSpecialtyRefDOMapper.updateByPrimaryKeySelective(hospSpecialtyRefDO);
     }
 
     /*删除科室记录，包括科室表和关系表*/
     @Override
-    public void deleteSpecialty(SpecialtyDOKey specialtyDOKey, HospSpecialtyRefDOKey hospSpecialtyRefDOKey) {
+    public void deleteSpecialty(SpecialtyDto specialtyDto) {
+        ValidatorResult result = validator.validate(specialtyDto);
+        if(result.isHasErrors()){
+            throw new BusinessException(result.getErrMsg(), EmBusinessError.PARAMETER_VALIDATION_ERROR);
+        }
+        SpecialtyDOKey specialtyDOKey = specialtyDto;
+        HospSpecialtyRefDOKey hospSpecialtyRefDOKey = new HospSpecialtyRefDOKey();
+        hospSpecialtyRefDOKey.setItemid(specialtyDto.getItemid());
+        hospSpecialtyRefDOKey.setItemcode(specialtyDto.getItemcode());
+
         hospSpecialtyRefDOMapper.deleteByPrimaryKey(hospSpecialtyRefDOKey);
         specialtyDOMapper.deleteByPrimaryKey(specialtyDOKey);
     }
@@ -59,6 +108,9 @@ public class SpecialtyServiceImpl implements ISpecialtyService {
      */
     @Override
     public List<SpecialtyDO> searchSpecialty(String keyWord) {
+        if(keyWord.isEmpty()){
+            throw new BusinessException("关键字不能为空", EmBusinessError.PARAMETER_VALIDATION_ERROR);
+        }
         return specialtyDOMapper.searchSpecialty(keyWord);
     }
 
@@ -67,6 +119,7 @@ public class SpecialtyServiceImpl implements ISpecialtyService {
     public List<SpecialtyDO> top5Specialty() {
         return specialtyDOMapper.top5Specialty();
     }
+
 
 
 }
